modifiy how imageUrl is saved

+++ src/service/places.ts
@@ -4,14 +4,12 @@ import { Injectable } from "@angular/core";
 import { NativeStorage } from "@ionic-native/native-storage";
 import { File } from "@ionic-native/file";
 
-declare var cordova: any;
 @Injectable()
 export class PlacesService {
     private places: Place[]=[];
 
     constructor(private storage: NativeStorage,
                 private file: File){
-
     }
     addPlace(title: string, 
             description: string,
@@ -22,7 +20,7 @@ export class PlacesService {
                 this.storage.setItem('places', this.places)
                     .then()
                     .catch(
-                        err=>{
+                        ()=>{
                             this.places.splice(this.places.indexOf(place),1);
                         }
                     );
@@ -51,18 +49,39 @@ export class PlacesService {
         this.places.splice(index,1);
         this.storage.setItem('places', this.places)
             .then(()=>{
-                this.removeFile(place);
+                if (place.imageUrl!==''){
+                    this.removeFile(place);
+                }
             })
-            .catch((err)=>console.log(err));
+            .catch((err)=>console.log(err.message));
     }
 
+    getDataUrl(index: number){
+        this.file.readAsDataURL(this.file.dataDirectory,this.places[index].imageUrl)
+            .then((res)=>{return res});
+    }
     private removeFile(place: Place){
-        const currentName=place.imageUrl.replace(/^.*[\\\/]/,'');
-        this.file.removeFile(cordova.file.dataDirectory,currentName)
-            .then(()=>console.log('removed file'))
-            .catch(()=>{
-                console.log('Error while removing File');
+        this.file.removeFile(this.file.dataDirectory,place.imageUrl)
+            .then(()=>alert('removed file'))
+            .catch((err)=>{
+                alert(err.message);
+                alert(place.imageUrl);
                 this.addPlace(place.title,place.description,place.location,place.imageUrl);
             });
+    }
+
+    listAll(){
+        this.file.listDir(this.file.dataDirectory,'files')
+            .then((files)=>{
+                for (let file of files){
+                    alert(file.name);
+                    // if (file.name.includes(".jpg")){
+                    //     file.remove(()=>{
+                    //         alert("delete"+file.name);
+                    //     },(err)=>{alert(err.message)});
+                    // }
+                }
+            })
+            .catch();
     }
 }
